refactor items overview to component
