+++ src/lib/types/item.d.ts
@@ -3,6 +3,8 @@ export type Item = {
   name: string
 }
 
+export type ItemUrlFn = (id: string, name: string) => string
+
 export type FetchItems = () => Promise<Item[]>
 
 export type RemoveItem = (itemId: string) => Promise
@@ -10,4 +12,6 @@ export type RemoveItem = (itemId: string) => Promise
 export type AddItem = (itemId: string) => Promise
 
 export type CreateItem = (name: string) => Promise<Item>
+
+export type SearchItems = (name: string) => Promse<Item[]>
